+++ CHANGES.md
@@ -1,6 +1,7 @@
 unreleased
 ----------
 
+- Escape newlines. (@MisterDA #192 #194, reported by @kit-ty-kate)
 - Add Ubuntu 23.10. (@MisterDA #189)
 - Deprecate Ubuntu 22.10 it is now EOL. (@tmcgilchrist #184).
 - Support `--start-interval` in `HEALTCHECK` Dockerfile instruction.